requirements.txt
alembic==1.13.2
fastapi==0.111.1
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
uvicorn==0.30.1
